[Refactor] Put image extension into ContentTableViewCell
