--- iOS Exercise/Extension/UIImageExtension.swift
@@ -1,23 +1,0 @@
-//
-//  UIImageExtension.swift
-//  iOS Exercise
-//
-//  Created by allen on 2022/1/17.
-//
-
-import UIKit
-
-extension UIImage {
-    func resizeTopAlignedToFill(newWidth: CGFloat) -> UIImage? {
-        let newHeight = size.height * newWidth / size.width
-
-        let newSize = CGSize(width: newWidth, height: newHeight)
-
-        UIGraphicsBeginImageContextWithOptions(newSize, false, UIScreen.main.scale)
-        draw(in: CGRect(origin: .zero, size: newSize))
-        let newImage = UIGraphicsGetImageFromCurrentImageContext()
-        UIGraphicsEndImageContext()
-
-        return newImage
-    }
-}
