+++ iOS Exercise/View/ContentTableViewCell.swift
@@ -8,6 +8,7 @@
 import UIKit
 
 class ContentTableViewCell: UITableViewCell {
+    
     @IBOutlet weak var plantImage: UIImageView!
     @IBOutlet weak var plantName: UILabel!
     @IBOutlet weak var plantLocation: UILabel!
@@ -25,5 +26,22 @@ class ContentTableViewCell: UITableViewCell {
         plantLocation.text = data.location
         plantFeature.text = data.feature
         plantImage.image = data.image?.resizeTopAlignedToFill(newWidth: 100) ?? nil
+    }
+}
+
+// MARK: - Private Extension for UIImage
+
+private extension UIImage {
+    func resizeTopAlignedToFill(newWidth: CGFloat) -> UIImage? {
+        let newHeight = size.height * newWidth / size.width
+
+        let newSize = CGSize(width: newWidth, height: newHeight)
+
+        UIGraphicsBeginImageContextWithOptions(newSize, false, UIScreen.main.scale)
+        draw(in: CGRect(origin: .zero, size: newSize))
+        let newImage = UIGraphicsGetImageFromCurrentImageContext()
+        UIGraphicsEndImageContext()
+
+        return newImage
     }
 }
